+++ core/Include/radar_cfar.h
@@ -17,18 +17,18 @@ typedef struct {
     int rangeDim;                   // 距离维大小
     int refCells[2];                // [速度维单侧参考单元数, 距离维单侧参考单元数]
     int guardCells[2];              // [速度维单侧保护单元数, 距离维单侧保护单元数]
-    double thresholdFactor;         // 线性检测门限参数（倍数）
+    float thresholdFactor;         // 线性检测门限参数（倍数）
 } CfarParams;
 
 // 定义检测结果结构体
 typedef struct {
     int rangeIdx;      // 距离维粗糙索引 (0-based)
-    double rangeFine;  // 距离维精细值 (Interpolation result)
+    float rangeFine;  // 距离维精细值 (Interpolation result)
     int velIdx;        // 速度维粗糙索引 (0-based)
-    double velFine;    // 速度维精细值 (Interpolation result)
-    double amplitude;  // 检测点的幅度
-    double snr;        // 信噪比 (dB)
-    double noise;      // 噪声估计值
+    float velFine;    // 速度维精细值 (Interpolation result)
+    float amplitude;  // 检测点的幅度
+    float snr;        // 信噪比 (dB)
+    float noise;      // 噪声估计值
 } DetectionInfo;
 
 extern CfarParams cfar_params; // 全局CFAR参数实例
@@ -58,10 +58,10 @@ int perform_cfar_detection(const RadarFFT2DOutput input_fft_2d_data,
  * @param fixedIdx 固定维度的索引（0基）。
  * @param isVelocityDim true=速度维插值，false=距离维插值。
  * @param dimSize 插值维度的总长度。
- * @return double 亚单元偏移量 delta。
+ * @return float 亚单元偏移量 delta。
  */
-static double peak_interpolation_core(
-    const double data_map[][RANGE_BINS],
+static float peak_interpolation_core(
+    const float data_map[][RANGE_BINS],
     int peakIdx, 
     int fixedIdx, 
     bool isVelocityDim, 
@@ -76,7 +76,7 @@ static double peak_interpolation_core(
  * @param rangeDim 距离维大小 (RANGE_BINS)。
  */
 void refine_detections_interpolation(
-    const double data_map[][RANGE_BINS],
+    const float data_map[][RANGE_BINS],
     DetectionInfo *detections,
     int count,
     int velDim,
